refactor: migrate to @if/@for control flow

+++ gedbrowserng-frontend/src/app/modules/person/person-family-child.component.ts
@@ -6,7 +6,6 @@ import { HasFamily } from '../../interfaces/has-family';
 import { HasPerson } from '../../interfaces/has-person';
 import { PersonGetter } from './person-getter';
 import { RefreshPerson } from '../../interfaces';
-import { NgIf } from '@angular/common';
 import { RouterLink } from '@angular/router';
 import { MatIconButton } from '@angular/material/button';
 import { MatTooltip } from '@angular/material/tooltip';
@@ -24,17 +23,21 @@ import { MatIcon } from '@angular/material/icon';
 @Component({
     selector: 'app-person-family-child',
     template: `<div class="parent">
-  <span *ngIf="person">
-    <a [routerLink]="['/' + dataset + '/persons', person.string]" class="name">{{ person.indexName }} {{ lifespanYearString() }} [{{ person.string }}]</a>
-  </span>
+    @if (person) {
+        <span>
+            <a [routerLink]="['/' + dataset + '/persons', person.string]" class="name">{{ person.indexName }} {{ lifespanYearString() }} [{{ person.string }}]</a>
+        </span>
+    }
   <span class="example-fill-remaining-space"></span>
-  <span class="hidden" *ngIf="hasSignedIn()">
-    <button mat-icon-button matTooltip="Unlink" color="warn" (click)="unlink()">
-      <mat-icon matListIcon>link_off</mat-icon></button>
-  </span>
+    @if (hasSignedIn()) {
+        <span class="hidden">
+            <button mat-icon-button matTooltip="Unlink" color="warn" (click)="unlink()">
+                <mat-icon matListIcon>link_off</mat-icon></button>
+        </span>
+    }
 </div>`,
     styles: [],
-    imports: [NgIf, RouterLink, MatIconButton, MatTooltip, MatIcon]
+    imports: [RouterLink, MatIconButton, MatTooltip, MatIcon]
 })
 export class PersonFamilyChildComponent extends PersonGetter
     implements OnInit, OnChanges, HasPerson {
