+++ gedbrowserng-frontend/src/app/modules/person/person-family.component.ts
@@ -7,7 +7,6 @@ import { InitablePersonCreator } from '../../bases';
 import { HasAttributeList, HasPerson, RefreshPerson, LinkCheck, Saveable } from '../../interfaces';
 import { MatCard, MatCardTitle, MatCardContent } from '@angular/material/card';
 import { MatToolbar } from '@angular/material/toolbar';
-import { NgIf } from '@angular/common';
 import { PersonFamilySpouseComponent } from './person-family-spouse.component';
 import { MatIconButton } from '@angular/material/button';
 import { MatTooltip } from '@angular/material/tooltip';
@@ -33,24 +32,33 @@ import { PersonFamilyChildListComponent } from './person-family-child-list.compo
     selector: 'app-person-family',
     template: `<mat-card>
   <mat-card-title>
-    <mat-toolbar color="primary">Family {{ index + 1 }}<span *ngIf="spouse()">&nbsp;-&nbsp;</span>
-      <app-person-family-spouse *ngIf="spouse()"
-          [dataset]="dataset" [parent]="this" [attribute]="spouse()"></app-person-family-spouse>
-      <span *ngIf="spouse() && hasSignedIn()">
-        <button mat-icon-button matTooltip="Unlink self from family" color="accent" (click)="unlink()">
-          <mat-icon matListIcon>link_off</mat-icon></button>
-      </span>
-      <span *ngIf="!spouse()" class="example-fill-remaining-space"></span>
-      <span *ngIf="!spouse() && hasSignedIn()">
-        <app-new-person
-            [sex]="sex" [surname]="surname" [label]="'Create spouse'"
-            (emitOK)="createPerson($event)"></app-new-person>
-        <app-link-person
-            [parent]="this" [dataset]="dataset" [multi]="false" [label]="'Link parent'"
-            (emitOK)="linkPerson($event)"></app-link-person>
-        <button mat-icon-button matTooltip="Unlink self from family" color="accent" (click)="unlink()">
-          <mat-icon matListIcon>link_off</mat-icon></button>
-      </span>
+        <mat-toolbar color="primary">Family {{ index + 1 }}
+            @if (spouse()) { <span>&nbsp;-&nbsp;</span> }
+            @if (spouse()) {
+                <app-person-family-spouse
+                        [dataset]="dataset" [parent]="this" [attribute]="spouse()"></app-person-family-spouse>
+            }
+            @if (spouse() && hasSignedIn()) {
+                <span>
+                    <button mat-icon-button matTooltip="Unlink self from family" color="accent" (click)="unlink()">
+                        <mat-icon matListIcon>link_off</mat-icon></button>
+                </span>
+            }
+            @if (!spouse()) {
+                <span class="example-fill-remaining-space"></span>
+            }
+            @if (!spouse() && hasSignedIn()) {
+                <span>
+                    <app-new-person
+                            [sex]="sex" [surname]="surname" [label]="'Create spouse'"
+                            (emitOK)="createPerson($event)"></app-new-person>
+                    <app-link-person
+                            [parent]="this" [dataset]="dataset" [multi]="false" [label]="'Link parent'"
+                            (emitOK)="linkPerson($event)"></app-link-person>
+                    <button mat-icon-button matTooltip="Unlink self from family" color="accent" (click)="unlink()">
+                        <mat-icon matListIcon>link_off</mat-icon></button>
+                </span>
+            }
     </mat-toolbar>
   </mat-card-title>
   <mat-card-content>
@@ -70,14 +78,16 @@ import { PersonFamilyChildListComponent } from './person-family-child-list.compo
     </div>
     <div class="ui-g-1"></div>
   </div>
-  <app-person-family-child-list *ngIf="family?.children"
-      [children]="family?.children" [family]="family" [parent]="this"
-      [dataset]="dataset"></app-person-family-child-list>
+    @if (family?.children) {
+        <app-person-family-child-list
+                [children]="family?.children" [family]="family" [parent]="this"
+                [dataset]="dataset"></app-person-family-child-list>
+    }
   </mat-card-content>
 </mat-card>
 <br/>`,
     styles: [],
-    imports: [MatCard, MatCardTitle, MatToolbar, NgIf, PersonFamilySpouseComponent, MatIconButton, MatTooltip, MatIcon, NewPersonComponent, LinkPersonComponent, MatCardContent, AttributeListComponent, MultimediaGalleryComponent, PersonFamilyChildListComponent]
+    imports: [MatCard, MatCardTitle, MatToolbar, PersonFamilySpouseComponent, MatIconButton, MatTooltip, MatIcon, NewPersonComponent, LinkPersonComponent, MatCardContent, AttributeListComponent, MultimediaGalleryComponent, PersonFamilyChildListComponent]
 })
 export class PersonFamilyComponent extends InitablePersonCreator
     implements HasAttributeList, LinkCheck, Saveable {
